Permit override of `name` for demo (#8)
My use case is to have multiple demos for a given component, where each demo showcases a different use case. Here, the assumption that the component name is the final element in the `location` array does not hold. For example:

```js
{
  location: ['MyButton', 'containing text'],
  description: 'Custom button with text',
  importPath: 'foo/MyButton'
}, {
  location: ['MyButton', 'containing emoji'],
  description: 'Custom button with emoji',
  importPath: 'foo/MyButton'
}
```

This leads to an incorrect import statement: `import containing text from 'foo/MyButton'`